# api/contracts/v1/api-specs.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger Userstore
  description: A sample API that uses a User as an example to demonstrate features in the OpenAPI 3.0 specification
  termsOfService: https://swagger.io/terms/
  contact:
    name: Go App API Team
    email: apiteam@swagger.io
    url: https://swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://Users.swagger.io/api
paths:
  /users:
    get:
      summary: List users with cursor-based pagination
      description: Returns a page of users ordered by createdAt ascending
      operationId: listUsers
      parameters:
        - name: after
          in: query
          description: Cursor — createdAt of the last item from the previous page (RFC 3339)
          required: false
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Maximum number of users to return (default 20, max 100)
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: paginated user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Creates a new user
      description: Creates a new user
      operationId: addUser
      requestBody:
        description: User to add to the store
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '200':
          description: user response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{id}:
    get:
      summary: Returns a User by ID
      description: Returns a User based on a single ID
      operationId: findUserByID
      parameters:
        - name: id
          in: path
          description: ID of User to fetch
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /openai/{topic}:
    get:
      summary: Returns a Paragraph
      description: Returns a Paragraph based on a topic
      operationId: getParagraphByTopic
      parameters:
        - name: topic
          in: path
          description: topic to search
          required: true
          schema:
            type: string
      responses:
        '200':
          description: open ai response
          content:
            application/json:
              schema:
                type: string
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'              
components:
  schemas:
    User:
      properties:
        firstName:
          type: string
        lastName:
          type: string
        mobile:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    NewUser:
      required:
        - firstName
        - email
      properties:
        firstName:
          type: string
        lastName:
          type: string
        mobile:
          type: string
        email:
          type: string

    UserPage:
      required:
        - users
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        nextCursor:
          type: string
          format: date-time
          description: Pass as `after` on the next request to get the next page; absent when no more pages exist

    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        message:
          type: string
          description: Error message